More verbose file migration logging
The log interval was useful to see errors in the noise,
but it's also masked what the task actually does.
Since this is moving and modifying project data,
it should be verbose by default. This will help answer
question later on, e.g. "why wasn't this file migrated".

A new message at the bottom which spells out how many files failed (if any)
makes it clear that you need to review the logs in detail (even though the error lines might hide between a lot of success lines).