FIX Force alt attribute to render, remove fallback to image Title
Some images are presentational, and therefore shouldn't have alt
content. Others have captions, and therefore the alt content would be
duplicate. Most importantly, the name of a file often doesn't accurately
or sufficiently describe its contents, and so it is unfit as alt
content.

To improve a11y, the alt attribute should always be rendered, even if
it's empty, and it should not have a default value.